server/requirements.txt
quart==0.19.9
asyncpg==0.29.0
python-dotenv==1.0.0
quart-schema==0.17.1
hypercorn==0.15.0